Signal Strength: Positive Signed Integer Domain
